define an initialzer rather than using before_initialize
the heroku_db_env railtie was building the database_configuration
*after* the 'active_record.initialize_database' which meant the database
connection was already established before the heroku_database.yaml could
be merged in. Secondly, the before_intialize code was running even when
config.assets.initialize_on_precompile was set to false. Defining an
initializer allows us to set the group as default which won't run on
asset precompile